Remove ModalBackdrop and create inrernal _ModalBackdrop
