--- Source/Blazorise.AntDesign/ModalBackdrop.razor
@@ -1,3 +1,0 @@
-﻿@inherits Blazorise.ModalBackdrop
-<div @ref="@ElementRef" id="@ElementId" class="@ClassNames" style="@StyleNames" @attributes="@Attributes"></div>
-@ChildContent
