--- Source/Blazorise/Components/Modal/ModalBackdrop.razor
@@ -1,5 +1,0 @@
-﻿@namespace Blazorise
-@inherits BaseComponent
-<div @ref="@ElementRef" id="@ElementId" class="@ClassNames" style="@StyleNames" @attributes="@Attributes">
-    @ChildContent
-</div>
