+++ Source/Blazorise/Components/Modal/_ModalBackdrop.razor
@@ -1,0 +1,3 @@
+﻿@namespace Blazorise
+@inherits BaseComponent
+<div @ref="@ElementRef" id="@ElementId" class="@ClassNames" style="@StyleNames" @attributes="@Attributes"></div>
